+++ src/L2Data2Code/L2Data2Code.SchemaReader/MySql/MySqlSchemaReader.cs
@@ -144,6 +144,7 @@ namespace L2Data2Code.SchemaReader.MySql
                 tbl.CleanName = RemoveTablePrefixes(nameResolver.ResolveTableName(tbl.Name)).PascalCamelCase(false);
                 tbl.Type = nameResolver.ResolveTableType(tbl.Name);
                 (tbl.EnumValue, tbl.EnumName) = nameResolver.ResolveEnumTables(tbl.Name);
+                (tbl.DescriptionId, tbl.DescriptionColumn) = nameResolver.ResolveDescriptionTables(tbl.Name);
                 tbl.ClassName = tbl.CleanName.ToSingular();
                 tbl.Description = alternativeDescriptions != null && alternativeDescriptions.TryGetValue(tbl.Name, out var value) ? value : (fromViews ? string.Empty : (string)row["TABLE_COMMENT"]);
                 tbl.IsWeakEntity = nameResolver.IsWeakEntity(tbl.Name);
@@ -186,7 +187,7 @@ namespace L2Data2Code.SchemaReader.MySql
                 {
                     col.DefaultValue += "m";
                 }
-                col.IsFilter = filteredColumns.Contains(col.Name);
+                TrySetFilterColumn(filteredColumns, col);
                 result.Add(col);
             }