+++ src/L2Data2Code/L2Data2Code.SchemaReader/SQLite/SQLiteSchemaReader.cs
@@ -140,6 +140,7 @@ namespace L2Data2Code.SchemaReader.MySql
                 tbl.CleanName = RemoveTablePrefixes(nameResolver.ResolveTableName(tbl.Name)).PascalCamelCase(false);
                 tbl.Type = nameResolver.ResolveTableType(tbl.Name);
                 (tbl.EnumValue, tbl.EnumName) = nameResolver.ResolveEnumTables(tbl.Name);
+                (tbl.DescriptionId, tbl.DescriptionColumn) = nameResolver.ResolveDescriptionTables(tbl.Name);
                 tbl.ClassName = tbl.CleanName.ToSingular();
                 tbl.Description = alternativeDescriptions != null && alternativeDescriptions.TryGetValue(tbl.Name, out var value) ? value : (fromViews ? (string)row["DESCRIPTION"] : string.Empty);
                 tbl.IsWeakEntity = nameResolver.IsWeakEntity(tbl.Name);
@@ -176,7 +177,7 @@ namespace L2Data2Code.SchemaReader.MySql
                 col.IsComputed = tbl.IsView;
                 col.DefaultValue = row["COLUMN_DEFAULT"].IfNull<string>(null) == null ? null : ((string)row["COLUMN_DEFAULT"]).RemoveOuter('(', ')').RemoveOuter('\'');
                 col.Description = alternativeDescriptions != null && alternativeDescriptions.TryGetValue(col.FullName, out var value) ? value : null;
-                col.IsFilter = filteredColumns.Contains(col.Name);
+                TrySetFilterColumn(filteredColumns, col);
                 result.Add(col);
             }