+++ src/L2Data2Code/L2Data2Code.SchemaReader/Schema/NameResolver.cs
@@ -1,6 +1,7 @@
 using L2Data2Code.SchemaReader.Configuration;
 using L2Data2Code.SchemaReader.Interface;
 using L2Data2Code.SharedLib.Configuration;
+using L2Data2Code.SharedLib.Extensions;
 using System.Collections.Generic;
 
 namespace L2Data2Code.SchemaReader.Schema
@@ -9,9 +10,10 @@ namespace L2Data2Code.SchemaReader.Schema
     {
         private Dictionary<string, string> tableNames = null;
         private Dictionary<string, string> columnNames = null;
-        private Dictionary<string , string> tableTypes = null;
-        private Dictionary<string , string> enumTables = null;
-        private Dictionary<string , List<string>> bigTables = null;
+        private Dictionary<string, string> tableTypes = null;
+        private Dictionary<string, string> enumTables = null;
+        private Dictionary<string, string> descriptionTables = null;
+        private Dictionary<string, Dictionary<string, ColumnFilter>> bigTables = null;
         private readonly List<string> weakEntities = [];
 
         private readonly IBasicConfiguration<SchemaConfiguration> schemas;
@@ -27,6 +29,7 @@ namespace L2Data2Code.SchemaReader.Schema
             columnNames = GetRenames(schemas[schemaName]?.RenameColumns);
             tableTypes = GetRenames(schemas[schemaName]?.TableTypes);
             enumTables = GetRenames(schemas[schemaName]?.EnumTables);
+            descriptionTables = GetRenames(schemas[schemaName]?.DescriptionTables);
             bigTables = GetBigTables(schemas[schemaName]?.BigTables);
             weakEntities.AddRange(GetSemiColonEntries(schemas[schemaName]?.WeakEntities));
         }
@@ -44,9 +47,21 @@ namespace L2Data2Code.SchemaReader.Schema
             tableTypes.TryGetValue(originalTableName, out var value) ? value : string.Empty;
 
         public (string id, string name) ResolveEnumTables(string originalTableName)
+            => GetDescriptionsPairs(originalTableName, enumTables);
+
+        public (string id, string name) ResolveDescriptionTables(string originalTableName)
+            => GetDescriptionsPairs(originalTableName, descriptionTables);
+
+        public bool IsWeakEntity(string originalTableName) => weakEntities.Contains(originalTableName);
+
+        public bool IsBigTable(string originalTableName) => bigTables.ContainsKey(originalTableName);
+
+        public Dictionary<string, ColumnFilter> GetBigTableColumns(string originalTableName) => bigTables.TryGetValue(originalTableName, out var value) ? value : [];
+
+        private (string id, string name) GetDescriptionsPairs(string originalTableName, Dictionary<string, string> descriptions)
         {
             (string id, string name) result = (null, null);
-            if (enumTables.TryGetValue(originalTableName, out var value) && value.Contains(','))
+            if (descriptions.TryGetValue(originalTableName, out var value) && value.Contains(','))
             {
                 var colums = value.Split(',');
                 result.id = colums[0];
@@ -54,12 +69,6 @@ namespace L2Data2Code.SchemaReader.Schema
             }
             return result;
         }
-
-        public bool IsWeakEntity(string originalTableName) => weakEntities.Contains(originalTableName);
-
-        public bool IsBigTable(string originalTableName) => bigTables.ContainsKey(originalTableName);
-
-        public List<string> GetBigTableColumns(string originalTableName) => bigTables.TryGetValue(originalTableName, out var value) ? value : [];
 
         private static Dictionary<string, string> GetRenames(string renameDescriptions)
         {
@@ -75,16 +84,33 @@ namespace L2Data2Code.SchemaReader.Schema
             return renames;
         }
 
-        private static Dictionary<string, List<string>> GetBigTables(List<BigTable> bigTables)
+        private static Dictionary<string, Dictionary<string, ColumnFilter>> GetBigTables(List<BigTable> bigTables)
         {
-            Dictionary<string, List<string>> result = new();
+            Dictionary<string, Dictionary<string, ColumnFilter>> result = new();
             if (bigTables != null)
             {
                 foreach (var bigTable in bigTables)
                 {
-                    result.Add(bigTable.Key, bigTable.ColumnsFilter);
+                    var columnsFilter = ExtractColumnFilter(bigTable.ColumnsFilter);
+                    result.Add(bigTable.Key, columnsFilter);
                 }
-            }   
+            }
+            return result;
+        }
+
+        private static Dictionary<string, ColumnFilter> ExtractColumnFilter(List<string> columnsFilter)
+        {
+            Dictionary<string, ColumnFilter> result = [];
+            if (columnsFilter != null)
+            {
+                foreach (var columnFilter in columnsFilter)
+                {
+                    var columnName = columnFilter.Piece(0, ':');
+                    var filterType = columnFilter.Piece(1, ':');
+                    var filterSpecification = columnFilter.Piece(2, ':');
+                    result.Add(columnName, new ColumnFilter { ColumnName = columnName, FilterType = filterType, FilterSpecification = filterSpecification });
+                }
+            }
             return result;
         }