+++ src/L2Data2Code/L2Data2Code.SchemaReader/Schema/SchemaReader.cs
@@ -1,5 +1,6 @@
 using L2Data2Code.SchemaReader.Interface;
 using L2Data2Code.SchemaReader.Lib;
+using System.Collections.Generic;
 
 namespace L2Data2Code.SchemaReader.Schema
 {
@@ -25,6 +26,22 @@ namespace L2Data2Code.SchemaReader.Schema
         public bool HasErrorMessage() => _outputWriter.ContainsErrorMessage;
 
         public virtual bool CanConnect() => true;
+
+        protected void TrySetFilterColumn(Dictionary<string, Configuration.ColumnFilter> filteredColumns, Column column)
+        {
+            if (filteredColumns.TryGetValue(column.Name, out var filter))
+            {
+                column.IsFilter = true;
+                column.FilterType = filter.FilterType;
+                column.FilterSpecification = filter.FilterSpecification;
+            }
+            else
+            {
+                column.IsFilter = false;
+                column.FilterType = null;
+                column.FilterSpecification = null;
+            }
+        }
 
     }