+++ src/L2Data2Code/L2Data2Code.SchemaReader/SqlServer/SqlServerSchemaReader.cs
@@ -77,6 +77,7 @@ namespace L2Data2Code.SchemaReader.SqlServer
                             tbl.CleanName = RemoveTablePrefixes(nameResolver.ResolveTableName(tbl.Name)).PascalCamelCase(false);
                             tbl.Type = nameResolver.ResolveTableType(tbl.Name);
                             (tbl.EnumValue, tbl.EnumName) = nameResolver.ResolveEnumTables(tbl.Name);
+                            (tbl.DescriptionId, tbl.DescriptionColumn) = nameResolver.ResolveDescriptionTables(tbl.Name);
                             tbl.ClassName = tbl.CleanName.ToSingular();
                             tbl.Description = columnsDescriptions.TryGetValue(tbl.Name, out var value) ? value
                                 : options.AlternativeDescriptions.TryGetValue(tbl.Name, out var alternativeValue) ? alternativeValue : null;
@@ -109,6 +110,7 @@ namespace L2Data2Code.SchemaReader.SqlServer
                             tbl.CleanName = RemoveTablePrefixes(nameResolver.ResolveTableName(tbl.Name)).PascalCamelCase(false);
                             tbl.Type = nameResolver.ResolveTableType(tbl.Name);
                             (tbl.EnumValue, tbl.EnumName) = nameResolver.ResolveEnumTables(tbl.Name);
+                            (tbl.DescriptionId, tbl.DescriptionColumn) = nameResolver.ResolveDescriptionTables(tbl.Name);
                             tbl.ClassName = tbl.CleanName.ToSingular();
                             tbl.Description = columnsDescriptions.TryGetValue(tbl.Name, out var value) ? value : null;
                             tbl.IsWeakEntity = nameResolver.IsWeakEntity(tbl.Name);
@@ -243,7 +245,7 @@ namespace L2Data2Code.SchemaReader.SqlServer
                     {
                         col.DefaultValue += "m";
                     }
-                    col.IsFilter = filteredColumns.Contains(col.Name);
+                    TrySetFilterColumn(filteredColumns, col);
 
                     result.Add(col);
                 }